Fix wasm build * Add `utxo_common_tests.rs`
